Oops.  This was missing from my earlier checkin.  It shouldn't have caused a problem.
This takes care of removing the deg symbol (°) and tick marks for minutes and seconds (' and ") from input to the LatLonConverter.
It also fixed a problem with stripping zeros from a number like 0.620000.  Previously the leading 0 was also stripped.  That's pretty much never what you want.  I changed a call to StringUtils.strip() to StringUtils.stripEnd().

git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@1856 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39