Fixed some tree viewer problems: 	sync button was enabled even when only one view was showing 	delete and add context menu items were enabled without checking the business rules 	delete button was always disabled b/c the business rules canDelete() method had not been implemented
Fixed the warning that was appearing at startup in relation to the short term cache not finding the mapping file.  This was caused by the calling code specifying a mapping file name, but not saving the mapping file at shutdown.  Now the short-term cache doesn't have a mapping file.  It is also cleared everytime the app shutsdown.

git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@2338 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39