Added popups when the user tries to do something to the tree def that would require changes to tree nodes.
For example:
If the user edits a tree level and changes anything that would require updating of a lot of "fullname" fields, a popup is shown and the action is stopped.
If the user tries to enforce a tree level, but nodes in the tree violate that rule, a popup is shown and the action is blocked.

git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@2636 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39