Fix for getting the sample DB builder working on Derby.
I tried running the sample DB builder on Derby, which I hadn't done in quite a while.  It didn't work.  I got an exception about too many fields being in the ORDER BY or GROUP BY clause of a SELECT statement that Hibernate generates.  Here is the offending line of code.

List<?> taxa       = HibernateUtil.getCurrentSession().createCriteria(Taxon.class).list();

(I think this got added during Rod's rework of the sample DB builder, allowing for incremental commits to the DB instead of one huge one at the end of the whole process.)

The fix...

List<?> taxa       = HibernateUtil.getCurrentSession().createQuery("select t from Taxon t").list();

In general, if you can use a straight Query, do it.  Criteria searches are not supported in the JPA.  In this case, they also seem to make more complex SELECT statements.  In this particular instance, none of the power of the Criteria system is really being used anyway.  We just needed ALL of the objects from the given table.

git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@2919 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39