This should have been included with the last checkin.  Here's the log message from that one.
"Fixed bugz# 4960.  The problem was caused by the order of the MouseListeners that were attached to the JList.  The "internal" one that updates list item selection needs to fire after my listener that determines if a click was on text or not.  If the click isn't on text, I set a flag so selection change requests are ignored."

git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@3086 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39