Add truncateBigDecimal Scriptlet
Formats a BigDecimal to how it would appear in Query Builder Results and
on Forms

Should hopefully provide enough flexibility for working with
BigDecimals' alongside the formatBigDecimal Scriplet in Reports and Labels

(NOTE: This and other Scriptlets return a String becuase those are not
formatted [have a 'pattern'] by IReport.
IReport formats all numeric values by default to only have two places after
the decimal unless specified in the JRXML).